Bug fixes, refactoring
The solver had been forced to find solutions <20 moves, to fix this maxDepth is now set to int.MaxValue.

All functions to do with calculating ids for cubes have been moved to a seperate class.

Phase 1 and 2 repeat a lot of code so they are moved into a single ida* function.